Django/Practice/requirements.txt
asgiref==3.6.0
Django==3.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
